Fixes basic compatibility with Steamworks SDK 162, and fixes access violation error.
This is a janky fix, but it gets this working again with SDK 162.

First change is an early return after checking isInitSuccess. Not ideal, but it prevents the access violation error when the appid is not present in the persons library.

Other change is to comment out deprecated RequestCurrentStats(); call. Unclear to me if line 1066 should remain.